assessment corrected

+++ gbdt_uncertainty/assessment.py
@@ -88,7 +88,7 @@ def prr_regression(targets, preds, measure, pos_label=1):
 
 def ood_detect(domain_labels, in_measure, out_measure, mode, pos_label=1):
     scores = np.concatenate((in_measure, out_measure), axis=0)
-    scores = np.asarray(scores, dtype=np.float128)
+    scores = np.asarray(scores, dtype=np.float64)
     if pos_label != 1:
         scores *= -1.0
